git: show vim commands when editing commit messages
Starting vim with Insert mode (startinsert) messes up a lot of defaults.

Among them, when going into Normal mode, vim doesn't display the command
as you type it. Unless it's an ex command (like :x or :cq).

With `set showcmd` on, it's displayed on the far right corner (not
ideal), but that's still better than nothing.