zsh: don't highlight on the first token
A long-running paint point, zsh-syntax-highlighter paints the first
token (usually the command) in red as it's typed. Not cool.

This commit overrides this behavior and will only warn after the first
token.

I tried to write the style as in (regex, 'none') or (regex, 'fg=default')
as indicated [here](https://zsh.sourceforge.io/Doc/Release/Zsh-Line-Editor.html#Character-Highlighting)
but it refuses to work. I guess the zsh-syntax-highlighting middleware
obstructs this.

So I settled for a dummy 'bg=black' that does the trick. It seems like
only one highlighter rule may apply to any one region, and patterns have
priority over tokens.

The `rm -rf` pattern was the example for this feature and it's nice to
have.