zsh: track zshenv
Last time I installed pyenv, it had changed the way it hooks into your
shell, doing the $PATH-setting part of the init on .zprofile (or .zshenv)
and the interactive part (completion and more) on .zshrc.

This commit brings .zshenv into the fold and moves the init of every
package/version manager there. Other dotfiles I follow like mcornella's
also do this.

The rationale being that, as this is read first thing for any shell
(including login and non-tty), environment variables are set here. In
particular, additions to $PATH.

References:
https://github.com/pyenv/pyenv/blob/0915c5b7b6b56b20e0de48edcc3ab3b94f33bbb8/README.md#basic-github-checkout
https://github.com/pyenv/pyenv/blob/0915c5b7b6b56b20e0de48edcc3ab3b94f33bbb8/README.md#advanced-configuration
https://unix.stackexchange.com/questions/71253/what-should-shouldnt-go-in-zshenv-zshrc-zlogin-zprofile-zlogout