zsh: mv functions, plugin-patches into OMZ-custom
Every .zsh file in the custom ohmyzsh dir is automatically loaded.

I don't know what I was thinking when I wrote plugin patches into a
plugin itself... It's better to have it outside, more accessible.
Although it will make the blame harder to navigate.

I considered moving most of .zshrc into functions.zsh, and maybe
aliases.zsh. But it's not worth it, I'm ok with things as they are.