zsh, gnome-terminal: tweak snazzy colors
I was ticked off by how some of the pseudo-snazzy colors in the p10k
prompt differed quite a bit from the actual snazzy colors, so I've
tried some alternatives and changed the yellow one.

Pseudo snazzy green was the one that sent me down this rabbit hole but
in the end I'll keep the current one... But I've commented a couple
alternatives that are closer in hue to the actual snazzy green, although
with different H/V.

I've also taken the chance to remove the unnecessary quotes around color
codes, they're strings without special characters.

Lastly, I changed the white color in gnome-terminal/snazzy because it
was too close to bright_white (default text color) and that messes with
zsh-autosuggestions (no color difference between typed and suggested
text).