zsh: replace nvm for nodenv
Motivation: zsh startup time is noticeable, enough time for me to type a
few characters. nvm was 65-80% of that startup time, so I've switched to
nodenv.

nodenv is like rbenv or pyenv, a tool that "does one thing and does it
well." nvm is like rvm, which does a lot of things and is a full-blown
piece of software that logically takes time to load.

References:
https://stevenvanbael.com/profiling-zsh-startup
$ man zsh modules
https://github.com/nodenv/nodenv#basic-github-checkout
https://github.com/nodenv/nodenv-installer#nodenv-installer
https://github.com/nodenv/nodenv/wiki/Why-nodenv%3F
https://github.com/rbenv/rbenv/wiki/Why-rbenv%3F