git: fix logb paging
Bat pipes to less by default, but it's not picking up the LESS=R
environment. It's not usually noticeable because git exports LESS=FRX.

But when the log's short enough it falls on the path of less'
quit-if-one-screen (-F), which I don't like.

So I fall back to the paging setup I use with delta.