zsh: fix `help vscode` properly
The previous diagnosis was wrong. help() wasn't picking up vscode
not because it's an alias declared after it. It's because "vscode" (a
string) is passed as an argument to the function and aliases don't
expand when a function is executed, only when it's read. [1]

So I turned vscode into a function, with the sole caveat being this
$vscode_path variable lying around in the environment. No big deal.

[1] https://stackoverflow.com/a/14527916