bat: tweak colors for command help messages
Option arguments in cyan is not cool cause it's similar to option's
green but brighter. So it draws attention away from options, which are
the important component of that association.

I tried blue but it presents the same problem, it's just Srcery's green
is too dim. So let's go with magenta, which is a big departure from
green and I'd say not as bright as cyan/blue.

On the other hand, end of options (--) can use a brighter color than
bright blue, cause it's not conspicuous next to options. Dashes take up
too little space next to letters.