tilix: add buffered changes
Brings back cursor color, because when it's unset then it takes the
color of the element its on. Which makes it hard to detect when in vim.

"Auto hide the mouse when typing" on cause it seems like a cool feature.

"Warn when attempting unsafe paste" off cause it complains when I copy
sudo commands to install things and it's a bother.