git, bat: properly scope refs in commit header
Adds a custom syntax for git log, based on this one in Sublime/Packages.
[1]

But the latest commit there [2], other than adding support for refs,
also breaks the scoping of the rest of the commit header & message.

Or perhaps "in my local", or on syntect/bat. Because it's supposed to
pass the syntax tests in Sublime/Packages.

[1] https://github.com/sublimehq/Packages/blob/653edd49aa3370f027be1c75177e24aaa781e3ac/Git%20Formats/Git%20Log.sublime-syntax
[2] https://github.com/sublimehq/Packages/commit/653edd49aa3370f027be1c75177e24aaa781e3ac