vim: don't put x-deleted chars in a register
Pain point: I often go from this

abcde, fghij

to this:

fghij, abcde

The usual way I go about it is daw on abcde, then delete what's in
between, reproduce it manually at the end of fghij and then p what was
originally deleted.

But I cannot delete the middle with x's because it stores the deleted
character into the default register, which forces me to enter Insert
mode just to hit Del or Backspace.

This commit changes the default behavior of x because I can't think of a
compelling use case where I'd want to keep a single deleted character.

I could make this a recursive mapping (noremap means "no recursive map")
but I'm not sure I'll ever need that and I'd rather avoid the potential
unintended consequences.

Based on some of the answers here:
https://stackoverflow.com/questions/54255/in-vim-is-there-a-way-to-delete-without-putting-text-in-the-register