zsh: don't hide exit status on help()
e.g.: $ help git stash push

`git stash push --help` fails but prints usage on stderr. Because help()
redirects 2>&1 and pipes that to bat, it prints that usage message and
one would think that's all there is to the help message for that
command.

But that's wrong, the `git stash push --help` command failed. There's no
help for the push sub-subcommand, only for git stash.

In short, the error status on the first command of the pipe is useful
information. This change keeps it.