git: improve commid UX
To start the commid message from the draft, the -F (file) option is much
better than the -t (template) one.

Oftentimes the drafted version is the final version, but `git commit -t`
complains when you don't change the message.

I'd always gotten around this by adding an x char to be x-deleted (vim).

But it works without a hitch with -F and --edit.

I don't know from which StackOverflow question did I get the idea to do
this with -t ...