vscode, p10k: use theme colors in terminal prompt
Tweak powerlevel10k config so it uses the terminal's 16 custom colors
when it's running in a vscode terminal. This way, it picks up the
theme-defined colors for the terminal.

Also, overhaul the color variables because they were a mess.

I'm unhappy with how dark_gray looks on the Lucario theme bg, not nearly
as off-black as it looks in srcery or lovelace-neon. But I can't up and
optimize it for Lucario when in vscode, because I also use other themes.

Credit:
https://stackoverflow.com/questions/57788150/can-i-detect-in-powershell-that-i-am-running-in-vs-codes-integrated-terminal