git: add delta navigation + better merge conflicts
Had to remove the piping to less, cause it breaks navigation (which
itself pipes to less with --pattern to match hunk, file and commit
boundaries).

But LESS=R works without piping to less too, I don't know why I had kept
that in the beginning (nothing in the blame).

merge.conflictstyle = diff3 is untested, but delta docs encourage
setting this so let's.

Also, updated delta for the first time since forever?

0.10.x to 0.15.0 just released (with improvement in within-line diffs!).

There's this weird error where it exits with a pipefail error if I
opened a big diff (say show -20) and I didn't reach the end of the diff.

It still happens if I revert the current changes. But I've not tested if
it happened with the previous delta version. Maybe this has always been
the case but I didn't notice cause I usually reach the end of diffs?